Create ShuffleArrayWithRandomSwaps
This function performs a shuffle on an integer array by randomly swapping the values of two different indices 30 times. If the same index is chosen twice in a single attempt, the iteration is skipped and not counted. After completing the 30 successful swaps, the function prints each value of the shuffled array on a new line. This simulates a randomized scrambling of the array's contents.