Create checkIfOdd
Receives a binary tree of integers and returns the value of the odd number in the tree if one exists; if all numbers are even, returns 0. Assumes there is at most one odd number in the tree. Internally uses a helper recursive function (returnOdd) to traverse the tree and find the odd value.