Create DogSitterSocialNetworkClasses
This code defines three classes to model users in a dog-sitting social network. The base class User manages a username with an option to change it. The Doggysitter subclass adds a price per hour and extends the change method to update both username and price. The Client subclass adds a dog’s name and allows changing both the dog’s name and username. The Main method demonstrates creating and updating instances interactively.